back-end/requirements.txt
greenlet==1.1.3
flask==2.1.2
flask-sqlalchemy==2.5.1
psycopg2-binary==2.9.3
pytest==7.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-Cors==3.0.9
requests==2.25.0
flask_bcrypt==1.0.1
flask_login==0.6.2
pyjwt==2.4.0